requirements.txt
PySide6>=6.7,<7.0
requests>=2.32,<3.0
python-dotenv>=1.0,<2.0
spotipy>=2.24,<3.0
google-api-python-client>=2.169,<3.0
google-auth>=2.40,<3.0
google-auth-oauthlib>=1.2,<2.0

